v4-bridge/v4-inbox-applier/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-json-logger==2.0.7
